renamed stuff and split into modules
